refactor(repo): add shared lib package for reusable utilities
