+++ apps/backend/package.json
@@ -23,6 +23,7 @@
   "dependencies": {
     "@yosemite-crew/auth": "workspace:^",
     "@yosemite-crew/database": "workspace:^",
+    "@yosemite-crew/lib": "workspace:^",
     "@prisma/client": "^5.22.0",
     "@aws-sdk/client-cognito-identity-provider": "3.772.0",
     "@aws-sdk/client-pinpoint": "^3.806.0",
