+++ packages/lib/package.json
@@ -1,0 +1,29 @@
+{
+  "name": "@yosemite-crew/lib",
+  "type": "module",
+  "main": "./dist/index.js",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
+  },
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc -p tsconfig.build.json",
+    "clean": "rm -rf dist",
+    "type-check": "tsc --noEmit"
+  },
+  "dependencies": {
+    "axios": "^1.15.0",
+    "validator": "^13.15.22"
+  },
+  "devDependencies": {
+    "@types/node": "^22.15.21",
+    "@types/validator": "^13.15.2",
+    "typescript": "5.0.4"
+  }
+}
