+++ packages/lib/src/types/common.ts
@@ -1,0 +1,5 @@
+export type Nullable<T> = T | null;
+export type Optional<T> = T | undefined;
+export type JsonPrimitive = string | number | boolean | null;
+export type JsonValue = JsonPrimitive | JsonObject | JsonValue[];
+export type JsonObject = { [key: string]: JsonValue };
